Add more (unnesserary) features
This is an extended version of the "I Love You" program in C++ that includes several new features designed to add personality and interactivity to the program.

In this version of the program, we've added two new random number generators that express just how much the user is loved compared to puppies and kittens. These numbers are generated at runtime and displayed to the user in a personalized message.

We've also updated the heart graphic to include more hearts and a slightly different shape, which we believe makes the graphic more attractive and engaging.

Finally, we've added a new personalized message at the end of the program thanking the user for being in the programmer's life. This message is designed to make the program feel more personal and to show the user that they are appreciated.

Overall, these changes are intended to make the "I Love You" program more engaging, more interactive, and more personal. We hope that users will enjoy these new features and that they will find the program to be a charming and fun expression of affection.